refactor: change package structure at domain layer
