+++ src/main/kotlin/com/tsundokuplus/application/service/UserService.kt
@@ -1,6 +1,6 @@
 package com.tsundokuplus.application.service
 
-import com.tsundokuplus.domain.model.User
+import com.tsundokuplus.domain.model.user.User
 import com.tsundokuplus.domain.repository.UserRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
