--- src/main/kotlin/com/tsundokuplus/domain/model/RoleType.kt
@@ -1,3 +1,0 @@
-package com.tsundokuplus.domain.model
-
-enum class RoleType { ADMIN, USER }
