+++ src/main/kotlin/com/tsundokuplus/domain/model/user/Email.kt
@@ -1,4 +1,4 @@
-package com.tsundokuplus.domain.model
+package com.tsundokuplus.domain.model.user
 
 @JvmInline
 value class Email(val value: String) {
