+++ src/main/kotlin/com/tsundokuplus/domain/model/user/RoleType.kt
@@ -1,0 +1,3 @@
+package com.tsundokuplus.domain.model.user
+
+enum class RoleType { ADMIN, USER }
