+++ src/main/kotlin/com/tsundokuplus/domain/repository/BookRepository.kt
@@ -1,6 +1,6 @@
 package com.tsundokuplus.domain.repository
 
-import com.tsundokuplus.domain.model.Book
+import com.tsundokuplus.domain.model.book.Book
 
 interface BookRepository {
     fun findAll(userId: Int): List<Book>
