+++ src/main/kotlin/com/tsundokuplus/domain/repository/UserRepository.kt
@@ -1,6 +1,6 @@
 package com.tsundokuplus.domain.repository
 
-import com.tsundokuplus.domain.model.User
+import com.tsundokuplus.domain.model.user.User
 
 interface UserRepository {
     fun findOne(email: String): User?
