+++ src/test/kotlin/com/tsundokuplus/application/controller/BookControllerTest.kt
@@ -5,7 +5,12 @@ import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import com.tsundokuplus.application.service.BookService
 import com.tsundokuplus.application.service.security.LoginUser
-import com.tsundokuplus.domain.model.*
+import com.tsundokuplus.domain.model.book.Book
+import com.tsundokuplus.domain.model.book.Note
+import com.tsundokuplus.domain.model.user.Email
+import com.tsundokuplus.domain.model.user.Password
+import com.tsundokuplus.domain.model.user.RoleType
+import com.tsundokuplus.domain.model.user.User
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
