+++ src/test/kotlin/com/tsundokuplus/domain/model/book/BookTest.kt
@@ -1,4 +1,4 @@
-package com.tsundokuplus.domain.model
+package com.tsundokuplus.domain.model.book
 
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
